+++ ios/Runner/Extensions/Bundle+Properties.swift
@@ -2,7 +2,7 @@
 //  Bundle+Properties.swift
 //  Runner
 //
-//  Created by Hiddify on 12/26/23.
+//  Created by Tomato on 12/26/23.
 //
 
 import Foundation