+++ ios/Runner/VPN/VPNManager.swift
@@ -102,7 +102,7 @@ class VPNManager: ObservableObject {
             `protocol`.providerBundleIdentifier = Bundle.main.baseBundleIdentifier + ".SingBoxPacketTunnel"
             `protocol`.serverAddress = "localhost"
             newManager.protocolConfiguration = `protocol`
-            newManager.localizedDescription = "Hiddify"
+            newManager.localizedDescription = "Tomato"
             try await newManager.saveToPreferences()
             try await newManager.loadFromPreferences()
             self.manager = newManager