+++ lib/features/window/notifier/window_notifier.dart
@@ -22,7 +22,7 @@ class WindowNotifier extends _$WindowNotifier with AppLogger {
 
     // if (Platform.isWindows) {
     //   loggy.debug("ensuring single instance");
-    //   await WindowsSingleInstance.ensureSingleInstance([], "Hiddify");
+    //   await WindowsSingleInstance.ensureSingleInstance([], "Tomato");
     // }
 
     await windowManager.ensureInitialized();
@@ -47,7 +47,11 @@ class WindowNotifier extends _$WindowNotifier with AppLogger {
   }
 
   Future<void> quit() async {
-    await ref.read(connectionNotifierProvider.notifier).abortConnection().timeout(const Duration(seconds: 2)).catchError(
+    await ref
+        .read(connectionNotifierProvider.notifier)
+        .abortConnection()
+        .timeout(const Duration(seconds: 2))
+        .catchError(
       (e) {
         loggy.warning("error aborting connection on quit", e);
       },