+++ linux/my_application.cc
@@ -47,14 +47,14 @@ static void my_application_activate(GApplication *application)
   {
     GtkHeaderBar *header_bar = GTK_HEADER_BAR(gtk_header_bar_new());
     gtk_widget_show(GTK_WIDGET(header_bar));
-    gtk_header_bar_set_title(header_bar, "Hiddify");
+    gtk_header_bar_set_title(header_bar, "Tomato");
     gtk_header_bar_set_show_close_button(header_bar, TRUE);
     gtk_window_set_titlebar(window, GTK_WIDGET(header_bar));
     
   }
   else
   {
-    gtk_window_set_title(window, "Hiddify");
+    gtk_window_set_title(window, "Tomato");
   }
 
   gtk_window_set_default_size(window, 1280, 720);